Extract date conversions into testable class.
